+++ fabric/src/main/java/com/chefmooon/colourfulclocks/common/util/fabric/ColourfulClocksItemPropertiesImpl.java
@@ -19,25 +19,8 @@ import org.jetbrains.annotations.Nullable;
 public class ColourfulClocksItemPropertiesImpl {
 
     public static void addCustomItemProperties() {
-        registerPocketWatch(ColourfulClocksItemsImpl.IRON_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.EXPOSED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.WEATHERED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.OXIDIZED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.WAXED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.WAXED_EXPOSED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.WAXED_WEATHERED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.WAXED_OXIDIZED_COPPER_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.GOLD_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.DIAMOND_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.NETHERITE_POCKET_WATCH.get());
-
-        registerPocketWatch(ColourfulClocksItemsImpl.EMERALD_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.AMETHYST_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.QUARTZ_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.LAPIS_LAZULI_POCKET_WATCH.get());
-        registerPocketWatch(ColourfulClocksItemsImpl.REDSTONE_POCKET_WATCH.get());
-
+        ColourfulClocksItemsImpl.POCKET_WATCH_VARIANTS.forEach(((pocketWatchTypes, itemSupplier) ->
+                registerPocketWatch(itemSupplier.get())));
         ColourfulClocksItemsImpl.BORNHOLM_MIDDLE_VARIANTS.forEach((entry, supplier) ->
                 registerBornholmMiddle(supplier.get()));
         ColourfulClocksItemsImpl.BORNHOLM_TOP_VARIANTS.forEach((entry, supplier) ->